Initial commit
This commit sets up 2 backends:
* url-service: responsible of returning the shorten URL
* key-manager-service: responsible of creating the 6 digits URL key

Both backends can be operated using Kubernetes inside a Minikube container.
They can talk to each other using a service mesh, Istio.

This project is meant to test Istio built-in features, the backends
implementation is straightforward.

This commit also provides a bunch a shell script to run and stop Minikube,
get the ingress gateway URL, build the project, and so on.

The current architecture is as below:
* Each backend lives in a pod alongside with the Istio sidecar.
* The sidecar talks to a backend using the port 8080 and exposes the backend
on the port 80.
* The url-service talks to the key-generator-service on the HTTP protocol.
* An ingress gateway exposes the HTTP URL `GET /url/create` outside of the
cluster.