Delete a URL from its hash
This commits adds the `/url/del?{hash}` endpoint.
Below is deltetion flow:
* First, try to remove the URL,
* Then, if the URL is removed, release the associated key (not used
anymore),
* Finally return a success, or an error if something went wrong.

The deletion flow isn't safe; if the URL is deleted but the key isn't,
then the key is lost forever. We could use an async queue to fix this
problem.

To achieve this, the key-generator-service uses the same Mongodb as the
url-service. This approach isn't service oriented but is fine for this
small project.
To cerate multiple keys, call the `/key/init` endpoint. Doing so will
generate and store 128 unique keys in the column `url.key` of the
database. 128 keys are enough for testing, but not for production. In a
real world we could generate all the possible keys and store them into
the database (~402 GB, too much for my computer).

Also, the error mapping could be improved, but once again, the goal is to
play with Istio.