Store admin user and password into kubernetes secrets
Secrets are meant to store sensible informations, such as database
account. This commit stores the mongodb admin account into a secret named
`mongodb-root-rw-secret` that contains the `MONGO_ROOT_USERNAME` and
`MONGO_ROOT_PASSWORD` key/values.
Also, the secret is stored in git which is unsafe, we should use sealed
secrets in the future.